fix(shared): split query server-only files under @langfuse/shared/query/server
