+++ packages/shared/package.json
@@ -50,6 +50,10 @@
     "./query": {
       "import": "./dist/src/features/query/index.js",
       "require": "./dist/src/features/query/index.js"
+    },
+    "./query/server": {
+      "import": "./dist/src/features/query/server/index.js",
+      "require": "./dist/src/features/query/server/index.js"
     }
   },
   "scripts": {
