+++ packages/shared/src/features/query/server/queryExecutor.ts
@@ -1,13 +1,13 @@
 import {
   queryClickhouse,
   type ClickhouseQueryOpts,
-} from "../../server/repositories/clickhouse";
-import { measureAndReturn } from "../../server/clickhouse/measureAndReturn";
-import { type PreferredClickhouseService } from "../../server/clickhouse/client";
+} from "../../../server/repositories/clickhouse";
+import { measureAndReturn } from "../../../server/clickhouse/measureAndReturn";
+import { type PreferredClickhouseService } from "../../../server/clickhouse/client";
 import { QueryBuilder } from "./queryBuilder";
-import { type QueryType, type ViewVersion } from "./types";
-import { getViewDeclaration } from "./dataModel";
-import { env } from "../../env";
+import { type QueryType, type ViewVersion } from "../types";
+import { getViewDeclaration } from "../dataModel";
+import { env } from "../../../env";
 
 export type PreparedQuery = {
   compiledQuery: string;
