+++ web/src/pages/api/public/v2/metrics.ts
@@ -7,7 +7,8 @@ import {
   GetMetricsV2Response,
 } from "@/src/features/public-api/types/metrics";
 import { InvalidRequestError, LangfuseNotFoundError } from "@langfuse/shared";
-import { executeQuery, validateQuery } from "@langfuse/shared/query";
+import { executeQuery } from "@langfuse/shared/query/server";
+import { validateQuery } from "@langfuse/shared/query";
 const DEFAULT_ROW_LIMIT = 100;
 
 export default withMiddlewares({
